Polishing, squashing bugs, testing via console
Adding main.py (just calls Battle function for now) in another commit, because I wanted to have these in a deeper folder.

Turns out that if you don't document your work the day you've done it, it may become difficult to keep track of specific changes.

Added testing battles via console text for now and random choices of moves. All of the features should work in this setting except for extinguishing status effects. Again, moves are randomized for now to speed up testing for bugs and balance. Overall, randomized battles generally end up being close for both parties - both the fire dragon and the shock snake fight head to head most of the time. This, of course, doesn't represent how balanced the game will be when moves start being used tactically in anticipation of the opponent or in combination with other moves, but it is a good sign.

Due to some bug fixes, features may be constricted: for example, you cannot instantly hear your enemy now - if such move is to be added in the future (i.e. heavy status affliction, but your enemy is instantly healed by X amount), there would need to be new boolean checks in place to allow that. With creatures currently in, everything works fine: statuses and ability cooldowns expire as they should, modifiers apply as needed, etc.

I'll add in the assets later - console (and I don't mean the PS5) is not the way it's meant to be played.