Button tips, ability cooldowns and animated roll
Also fix to crashing when mouse interacts with game window.

Button tips are for human players only and are in black. Info tip is still missing, but so is the functionality!
Ability cooldowns display in place of button tips. You don't need button tips for abilities on cooldown. They are light gray to allow distinguishing.
Animated rolling is introduced for hitting target and applying statuses. It's just text, but there are random rolls before the actual roll appears!

As far as I'm concerned, that's a pretty cool update.