"Who's first?", "Who's turn?" AI text, 6th ability
There is now information in text in GUI pre-round that shows why creature moves first - whether it had a faster move or got randomly selected. It's in the same spot as roll information.

Creature name's color now serves as information about who's turn it is (to differentiate who the text in textbox is about when there's two same creatures against one another) and whether it's gaining or losing health. It can be white-ish (not your turn) or black (your turn) and red or bleeding white (taking damage/taking damage on opponent's turn), and green or blooming white (healing/healing on opponent's turn).

There is now information in text in GUI showcasing AI's decisions on choosing a risky move or a counter-counter move. Basically this is to showcase AI features in decision making that are otherwise invisible (they were previously visible via console).

Added a universal 6th ability to all creatures - health kit. It's slow, heals quite a bit and removes all status effects. 12 round cooldown - fortunately it looks alright in the HUD despite being double digits. In modes with multiple creatures per player (if I get to adding that beast - I'm not sure if it's worth it for project this size, 3vs3 will take a while to play out, 1vs1 already does), I'll introduce switching creatures through that ability. I'll probably change it later so it goes on cooldown immediately when creature enters the arena. I'll probably update the graphic for it also as it's one of the types that I've previously introduced.

Polishing, mainly with how textboxes work, and bugfixes. I'm quite happy now with textbox timings and contents during all parts of the battle.