Human players, ready states, additional info
Added the ability for human players to play - both as player 1 and 2, simultaneously or not.

During the phase of choosing your move, player HUD will have a ring around it that's either green (READY), red (NOT READY) and blue (STUNNED) to give player feedback.
This design philosophy of not showing which move was chosen exactly (I thought about selecting with WSAD/arrows and Spacebar/ENTER previously) is to allow fair hot-seat player vs player battles. Ready states let the player know when input is expected and when it's been processed. Blue for stunned is to avoid confusion from going to green without pressing any buttons or from round starting from red. 

On top of the HUD, next to health bar and health text, additional information is displayed: aim bonus, defense bonus, thorn damage low, thorn damage high and base defense. This ensures quick access to strategically necessary information. This displays only in the phase of choosing your move when it's useful.

Balancing changes to Health Kit ability and Psywarca.

Added extinguishment scoring attribute to StatusOccurrence. Yes, this sort of thing logically should probably be in player.py as an array of modifiers instead like the previous creature move multiplier, but for the sake of simplicity I'm doing it this way. Anyhow, this attribute is added to make sure AI sees punishment in extinguishing good status effects with the Health Kit ability or any other ability that can extinguish status effects, and so it knows when extinguishing effects on opponent is valuable.