Info boxes part 1, improved implementation some
Implemented info boxes under TAB and NP_PLUS for Player 1 and Player 2.
Info boxes display information about player's or opponent's moves and player's or opponent's active status effects. 
Both players can browse information about each other separately.
They are not drawing ability and status icons yet, but the info box page is fully functioning and can be navigated with WASD QE and 5123 46.
The plan is to add that and add more pages that can be switched via Q/E or NP_4/NP_5, perhaps a type map or a creature list. 
In any way, this was the most crucial part of info buttons, the other ones are secondary and not completely necessary.
Part 2 will have the ability and status icons as well as some sort of way to tell which box has been selected other than reading text explaining ability/status.

Changed implementation of some features like keys to make the code cleaner. All of these changes are inside of gui.py.

Changed type of "TRIPPED" effect to "WIND". If you are immune to wind, you wouldn't trip from moves that have this effect, so it makes more sense this way.